Added movie fanart
Added movie fanart but it is not yet enabled as I do not have an api key which
I can distribute.  If you have your own api key you can put it in to help me
test the script.

I had to rearrange a lot of the code to accomodate this addition so there may
be unexpected bugs or regressions.