Feature: Centralized fanart
Adding the ability to place all fanart in a central location for each
media type to allow the user to take advantage of the multi-image
backdrops that can be enabled in some skins.  As of right now, I have
disabled the feature by default.  Because a directory outside of the
.xbmc directory is required, I have set $HOME/fanart/'mediatype' as the
default for each, but this can be changed in the add-on settings.  Also
gave us a version bump.