Master branch now works with both Eden and Dharma
With media and file functions separated, it made it a lot easier to
code functionality for both versions.  The script also runs just fine
in Dharma with xbmc.python set to 2.0.  File the file functions work
just fine for Dharma, we'll probably need to add more error checking
and logging in those functions.